+++ package.json
@@ -3,6 +3,7 @@
   "scripts": {
     "dev-client": "webpack-dev-server --mode=development --port 9000",
     "dev-server": "nodemon ./server/index.js localhost 3000",
+    "initDB": "node ./initDB/app.js",
     "build-client": "webpack --mode production",
     "test": "jest",
     "test-dev": "jest --watch",
